Calculate the modified CV from the CV

This function calculates the modified coefficient of variation (CV) based on a (unmodified) CV. The modified CV is calculated based on the rules in CMH-17-1G. Those rules are:

For CV < 4%, CV* = 6%

For 4% <= CV < 8%, CV* = CV / 2 + 4%

For CV > 8%, CV* = CV

calc_cv_star(cv)

Arguments

cv
The CV to modify

Value

The value of the modified CV

References

"Composite Materials Handbook, Volume 1. Polymer Matrix Composites Guideline for Characterization of Structural Materials," SAE International, CMH-17-1G, Mar. 2012.

See also

Examples

# The modified CV for values of CV smaller than 4% is 6%
calc_cv_star(0.01)

#> [1] 0.06

## [1] 0.06

# The modified CV for values of CV larger than 8% is unchanged
calc_cv_star(0.09)

#> [1] 0.09

## [1] 0.09