Test for decrease in mean or minimum individual

This test is used when determining if a new process or manufacturing location produces material properties that are "equivalent" to an existing dataset, and hence the existing basis values are applicable to the new dataset. This test is also sometimes used for determining if a new batch of material is acceptable. This function determines thresholds based on both minimum individual and mean, and optionally evaluates a sample against those thresholds. The joint distribution between the sample mean and sample minimum is used to generate these thresholds. When there is no true difference between the existing ("qualification") and the new population from which the sample is obtained, there is a probability of \(\alpha\) of falsely concluding that there is a difference in mean or variance. It is assumed that both the original and new populations are normally distributed. According to Vangel (2002), this test provides improved power compared with a test of mean and standard deviation.

equiv_mean_extremum(
  df_qual = NULL,
  data_qual = NULL,
  mean_qual = NULL,
  sd_qual = NULL,
  data_sample = NULL,
  n_sample = NULL,
  alpha,
  modcv = FALSE
)

Arguments

df_qual
(optional) a data.frame containing the qualification data. Defaults to NULL.

data_qual
(optional) a vector of observations from the "qualification" data to which equivalency is being tested. Or the column of df_qual that contains this data. Defaults to NULL

mean_qual
(optional) the mean from the "qualification" data to which equivalency is being tested. Defaults to NULL

sd_qual
(optional) the standard deviation from the "qualification" data to which equivalency is being tested. Defaults to NULL

data_sample
(optional) a vector of observations from the sample for which equivalency is being tested. Defaults to NULL

n_sample
(optional) the number of observations in the sample for which equivalency will be tested. Defaults to NULL

alpha
the acceptable probability of a type I error

modcv
(optional) a boolean value indicating whether a modified CV should be used. Defaults to FALSE, in which case the standard deviation supplied (or calculated from data_qual) will be used directly.

Value

Returns an object of class equiv_mean_extremum. This object is a list with the following named elements:

call

the expression used to call this function

alpha

the value of alpha passed to this function

n_sample

the number of observations in the sample for which equivalency is being checked. This is either the value n_sample passed to this function or the length of the vector data_sample.

k1

the factor used to calculate the minimum individual threshold. The minimum individual threshold is calculated as \(W_{min} = qual\,mean - k_1 \cdot qual\,sd\)

k2

the factor used to calculate the threshold for mean. The threshold for mean is calculated as \(W_{mean} = qual\,mean - k_2 \cdot qual\,sd\)

modcv

logical value indicating whether the acceptance thresholds are calculated using the modified CV approach

cv

the coefficient of variation of the qualification data. This value is not modified, even if modcv=TRUE

cv_star

The modified coefficient of variation. If modcv=FALSE, this will be NULL

threshold_min_indiv

The calculated threshold value for minimum individual

threshold_mean

The calculated threshold value for mean

result_min_indiv

a character vector of either "PASS" or "FAIL" indicating whether the data from data_sample passes the test for minimum individual. If data_sample was not supplied, this value will be NULL

result_mean

a character vector of either "PASS" or "FAIL" indicating whether the data from data_sample passes the test for mean. If data_sample was not supplied, this value will be NULL

min_sample

The minimum value from the vector data_sample. if data_sample was not supplied, this will have a value of NULL

mean_sample

The mean value from the vector data_sample. If data_sample was not supplied, this will have a value of NULL

Details

This function is used to determine acceptance limits for a sample mean and sample minimum. These acceptance limits are often used to set acceptance limits for material strength for each lot of material, or each new manufacturing site. When a sample meets the criteria that its mean and its minimum are both greater than these limits, then one may accept the lot of material or the new manufacturing site.

This procedure is used to ensure that the strength of material processed at a second site, or made with a new batch of material are not degraded relative to the data originally used to determine basis values for the material. For more information about the use of this procedure, see CMH-17-1G or PS-ACE 100-2002-006.

There are several optional arguments to this function. However, you can't omit all of the optional arguments. You must supply either data_sample or n_sample, but not both. You must also supply either data_qual (and df_qual if data_qual is a variable name and not a vector) or both mean_qual and sd_qual, but if you supply data_qual (and possibly df_qual) you should not supply either mean_qual or sd_qual (and visa-versa). This function will issue a warning or error if you violate any of these rules.

If modcv is TRUE, the standard deviation used to calculate the thresholds will be replaced with a standard deviation calculated using the Modified Coefficient of Variation (CV) approach. The Modified CV approach is a way of adding extra variance to the qualification data in the case that the qualification data has less variance than expected, which sometimes occurs when qualification testing is performed in a short period of time. Using the Modified CV approach, the standard deviation is calculated by multiplying CV_star * mean_qual where mean_qual is either the value supplied or the value calculated by mean(data_qual) and \(CV*\) is the value computed by calc_cv_star.

References

M. G. Vangel. Lot Acceptance and Compliance Testing Using the Sample Mean and an Extremum, Technometrics, vol. 44, no. 3. pp. 242–249. 2002.

“Composite Materials Handbook, Volume 1. Polymer Matrix Composites Guideline for Characterization of Structural Materials,” SAE International, CMH-17-1G, Mar. 2012.

Federal Aviation Administration, “Material Qualification and Equivalency for Polymer Matrix Composite Material Systems,” PS-ACE 100-2002-006, Sep. 2003.

See also

Examples

equiv_mean_extremum(alpha = 0.01, n_sample = 6,
                    mean_qual = 100, sd_qual = 5.5, modcv = TRUE)

#> 
#> Call:
#> equiv_mean_extremum(mean_qual = 100, sd_qual = 5.5, n_sample = 6, 
#>     alpha = 0.01, modcv = TRUE)
#> 
#> Modified CV used: CV* = 0.0675 ( CV = 0.055 )
#> 
#> For alpha = 0.01 and n = 6 
#> ( k1 = 3.128346 and k2 = 1.044342 )
#>                  Min Individual   Sample Mean   
#>      Thresholds:    78.88367        92.95069    

##
## Call:
## equiv_mean_extremum(mean_qual = 100, sd_qual = 5.5, n_sample = 6,
##     alpha = 0.01, modcv = TRUE)
##
## Modified CV used: CV* = 0.0675 ( CV = 0.055 )
##
## For alpha = 0.01 and n = 6
## ( k1 = 3.128346 and k2 = 1.044342 )
##                   Min Individual   Sample Mean
##      Thresholds:    78.88367        92.95069