Glance at an equiv_mean_extremum object

Glance accepts an object of type equiv_mean_extremum and returns a tibble::tibble() with one row of summaries.

Glance does not do any calculations: it just gathers the results in a tibble.

# S3 method for equiv_mean_extremum
glance(x, ...)

Arguments

x

an equiv_mean_extremum object returned from equiv_mean_extremum()

...

Additional arguments. Not used. Included only to match generic signature.

Value

A one-row tibble::tibble() with the following columns:

alpha the value of alpha passed to this function

n_sample the number of observations in the sample for which equivalency is being checked. This is either the value n_sample passed to this function or the length of the vector data_sample.

modcv logical value indicating whether the acceptance thresholds are calculated using the modified CV approach

threshold_min_indiv The calculated threshold value for minimum individual

threshold_mean The calculated threshold value for mean

result_min_indiv a character vector of either "PASS" or "FAIL" indicating whether the data from data_sample passes the test for minimum individual. If data_sample was not supplied, this value will be NULL

result_mean a character vector of either "PASS" or "FAIL" indicating whether the data from data_sample passes the test for mean. If data_sample was not supplied, this value will be NULL

min_sample The minimum value from the vector data_sample. if data_sample was not supplied, this will have a value of NULL

mean_sample The mean value from the vector data_sample. If data_sample was not supplied, this will have a value of NULL

See also

Examples

x0 <- rnorm(30, 100, 4)
x1 <- rnorm(5, 91, 7)
eq <- equiv_mean_extremum(data_qual = x0, data_sample = x1, alpha = 0.01)
glance(eq)
#> # A tibble: 1 × 9
#>   alpha n_sample modcv threshold_min_indiv threshold_mean result_min_indiv
#>   <dbl>    <int> <lgl>               <dbl>          <dbl> <chr>           
#> 1  0.01        5 FALSE                88.5           95.5 FAIL            
#> # ℹ 3 more variables: result_mean <chr>, min_sample <dbl>, mean_sample <dbl>

## # A tibble: 1 x 9
##   alpha n_sample modcv threshold_min_indiv threshold_mean
##   <dbl>    <int> <lgl>               <dbl>          <dbl>
## 1  0.01        5 FALSE                86.2           94.9
## # ... with 4 more variables: result_min_indiv <chr>, result_mean <chr>,
## #   min_sample <dbl>, mean_sample <dbl>