Glance at a mnr (maximum normed residual) object

Glance accepts an object of type mnr and returns a tibble::tibble() with one row of summaries.

Glance does not do any calculations: it just gathers the results in a tibble.

# S3 method for mnr
glance(x, ...)

Arguments

x
An mnr object

...
Additional arguments. Not used. Included only to match generic signature.

Value

A one-row tibble::tibble() with the following columns:

mnr the computed MNR test statistic

alpha the value of alpha used for the test

crit the critical value given the sample size and the significance level

n_outliers the number of outliers found

See also

Examples

x <- c(rnorm(20, 100, 5), 10)
m <- maximum_normed_residual(x = x)
glance(m)

#> # A tibble: 1 x 4
#>     mnr alpha  crit n_outliers
#>   <dbl> <dbl> <dbl>      <dbl>
#> 1  4.26  0.05  2.73          1

## # A tibble: 1 x 4
##     mnr alpha  crit n_outliers
##   <dbl> <dbl> <dbl>      <dbl>
## 1  4.23  0.05  2.73          1