Levene's Test for Equality of Variance

This function performs the Levene's test for equality of variance.

levene_test(data = NULL, x, groups, alpha = 0.05, modcv = FALSE)

Arguments

data
a data.frame

x
the variable in the data.frame or a vector on which to perform the Levene's test (usually strength)

groups
a variable in the data.frame that defines the groups

alpha
the significance level (default 0.05)

modcv
a logical value indicating whether the modified CV approach should be used.

Value

Returns an object of class adk. This object has the following fields:

call

the expression used to call this function

data

the original data supplied by the user

groups

a vector of the groups used in the computation

alpha

the value of alpha specified

modcv

a logical value indicating whether the modified CV approach was used.

n

the total number of observations

k

the number of groups

f

the value of the F test statistic

p

the computed p-value

reject_equal_variance

a boolean value indicating whether the null hypothesis that all samples have the same variance is rejected

modcv_transformed_data

the data after the modified CV transformation

Details

This function performs the Levene's test for equality of variance. The data is transformed as follows:

$$w_{ij} = \left| x_{ij} - m_i \right|$$

Where \(m_i\) is median of the \(ith\) group. An F-Test is then performed on the transformed data.

When modcv=TRUE, the data from each group is first transformed according to the modified coefficient of variation (CV) rules before performing Levene's test.

References

“Composite Materials Handbook, Volume 1. Polymer Matrix Composites Guideline for Characterization of Structural Materials,” SAE International, CMH-17-1G, Mar. 2012.

See also

Examples

library(dplyr)

carbon.fabric.2 %>%
  filter(test == "FC") %>%
  levene_test(strength, condition)

#> 
#> Call:
#> levene_test(data = ., x = strength, groups = condition)
#> 
#> n = 91           k = 5            
#> F = 3.883818     p-value = 0.00600518 
#> Conclusion: Samples have unequal variance ( alpha = 0.05 )
#> 

##
## Call:
## levene_test(data = ., x = strength, groups = condition)
##
## n = 91          k = 5
## F = 3.883818    p-value = 0.00600518
## Conclusion: Samples have unequal variance ( alpha = 0.05 )