Glance at a equiv_change_mean object

Glance accepts an object of type equiv_change_mean and returns a tibble::tibble with one row of summaries.

Glance does not do any calculations: it just gathers the results in a tibble.

# S3 method for equiv_change_mean
glance(x, ...)

Arguments

x
a equiv_change_mean object returned from equiv_change_mean

...
Additional arguments. Not used. Included only to match generic signature.

Value

A one-row tibble::tibble with the following columns:

alpha

the value of alpha passed to this function

n_sample

the number of observations in the sample for which equivalency is being checked. This is either the value n_sample passed to this function or the length of the vector data_sample.

mean_sample

the mean of the observations in the sample for which equivalency is being checked. This is either the value mean_sample passed to this function or the mean of the vector data-sample.

sd_sample

the standard deviation of the observations in the sample for which equivalency is being checked. This is either the value mean_sample passed to this function or the standard deviation of the vector data-sample.

n_qual

the number of observations in the qualification data to which the sample is being compared for equivalency. This is either the value n_qual passed to this function or the length of the vector data_qual.

mean_qual

the mean of the qualification data to which the sample is being compared for equivalency. This is either the value mean_qual passed to this function or the mean of the vector data_qual.

sd_qual

the standard deviation of the qualification data to which the sample is being compared for equivalency. This is either the value mean_qual passed to this function or the standard deviation of the vector data_qual.

modcv

logical value indicating whether the equivalency calculations were performed using the modified CV approach

sp

the value of the pooled standard deviation. If modecv = TRUE, this pooled standard deviation includes the modification to the qualification CV.

t0

the test statistic

t_req

the t-value for \(\alpha / 2\) and \(df = n1 + n2 -2\)

threshold_min

the minimum value of the sample mean that would result in a pass

threshold_max

the maximum value of the sample mean that would result in a pass

result

a character vector of either "PASS" or "FAIL" indicating the result of the test for change in mean

See also

Examples

x0 <- rnorm(30, 100, 4)
x1 <- rnorm(5, 91, 7)
eq <- equiv_change_mean(data_qual = x0, data_sample = x1, alpha = 0.01)
glance(eq)

#> # A tibble: 1 x 14
#>   alpha n_sample mean_sample sd_sample n_qual mean_qual sd_qual modcv    sp
#>   <dbl>    <int>       <dbl>     <dbl>  <int>     <dbl>   <dbl> <lgl> <dbl>
#> 1  0.01        5        90.7      7.25     30      99.5    3.88 FALSE  4.43
#> # … with 5 more variables: t0 <dbl>, t_req <dbl>, threshold_min <dbl>,
#> #   threshold_max <dbl>, result <chr>

## # A tibble: 1 x 14
##   alpha n_sample mean_sample sd_sample n_qual mean_qual sd_qual modcv
##   <dbl>    <int>       <dbl>     <dbl>  <int>     <dbl>   <dbl> <lgl>
## 1  0.01        5        85.8      9.93     30      100.    3.90 FALSE
## # ... with 6 more variables: sp <dbl>, t0 <dbl>, t_req <dbl>,
## #   threshold_min <dbl>, threshold_max <dbl>, result <chr>